pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dota_drafts"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Alhimovich Matvei", email = "matalhim@gmail.com" }]
dependencies = [
    "terminal_app @ git+https://github.com/usecoreai/terminal_app.git",
    "pytest",
    "pytest-mock",
    "typer",
    "watchdog",
    "python-dotenv",
    "httpx",
    "pydantic",
    "rich",
    "tqdm",
    "Pillow",
    "chromadb",
    "llama-index-core",
    "llama-index-readers-web",
    "llama-index-embeddings-huggingface",
    "llama-index-vector-stores-chroma",
    "llama-index-program-openai; python_version >= '3.11' and platform_system != 'Darwin'",
    "transformers",
    "sentence-transformers",
    "torch",
    "opencv-python-headless",
    "pytesseract",
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-text-splitters",
    "langsmith",
    "langchain-google-genai",
    "langchain-openai",
    "langchain-together",
    "langchain-huggingface",
    "streamlit",
    "st-img-pastebutton",
]
dynamic = ["version"]

# [project.urls]
# GitHub = "https://github.com/usecoreai/recommendation_system"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.dynamic]
version = { attr = "dota_drafts.__version__" }

[tool.setuptools]
include-package-data = true

# [tool.setuptools.package-data]
# recommendation_system = ["**/*.txt"]

[project.scripts]
dota-drafts = "dota_drafts.cli:app"
